Rework autostart and manager mode
Always show the app when sticky is run unless we're in autostart mode.

Hide notes by default on the screen in autostart mode.

Add a setting to show notes by default in autostart mode.

Add a setting for autostart and hide sticky in Startup Applications (this is to prevent people from duplicating the .desktop file).