note: Fix icons
We're on top of colored notes here, the icons don't need to adapt to the the rest of the
theme, they need to adapt to the notes themselves.

We can't use symbolic because they highlight when hovered.

We can't use white mono icons on dark themes, because the notes aren't dark.

We need to use our own icons instead to guarantee they always look the same.