Make sure the notes are visible before calling present_with_time()
If they aren't present_with_time() only shows them, but never raises them to the top or focuses them. This fixes the issue where the tray would need to be clicked twice to hide the notes.

This is the proper fix for 572c0bee95872c2ea1a0a97ee3a941d04bf37d34