Tray: fix notes getting 'stuck' in unraised state (I hope)
The problem seems to be some sort of weird interaction with the dummy window and either gtk or the window manager (I didn't try reproducing it in other window managers besides muffin). Calling `present_with_time()` on the dummy window before doing the same on the notes themselves seems to make it work properly (at least I haven't been able to reproduce it after adding that call, so if it can still happen, it's at least more rare which is a win).