Make the dummy window transparent
Despite our best efforts, sometimes the dummy window still ends up on-screen. It seems to be the window manager trying to be smart, and there doesn't seem to be a way to prevent this behavior. It is consistently reproducible in xfce, but seems to also affect Cinnamon in some situations, and likely other desktops as well.

Unfortunately, there doesn't seem to be a way to make the dummy window unreactive, but at least making it transparent means that it wont cause the annoying pixel on the screen. As for the reactivity, it is very unlikely that the user clicks exactly on the dummy window since it is only 1 pixel in size, so hopefully this fix is sufficient. If this does turn out to be an issue, we may have to go back to the drawing board.

Fixes: #31
Reference: #26