Note: cache the internal markup to increase performance on some updates.
This is probably most beneficial when dragging or editing notes while there is another large note on the screen. Previously, such actions would call buffer.get_internal_markup() (which is quite hefty for large notes), even if the text of that note didn't change. Now it's only called when needed.