buffer, do_insert_text: always reassign location text iter to former pos
As noted in the added comment, the TextBuffer.do_insert_text() virtual method shifts the location iter by one character. This lead to inconsistent results when it is reinitialized by offset in certain cases (such as with checkboxes and bullets).

This was causing at least one bug where the last 2 characters of a link weren't being highlighted when the link was in a line with a checkbox or bullet and the user presses enter.

By making this change, we now get consistent results, and we can now remove some workarounds that adjusted for this behavior.