Buffer: fix an issue where pasted text sometimes gets jumbled
The issue is that sometimes the pasted text gets split into pieces (eg. when pasting text from the same note which contains tag toggles). When that happens, the function that calls .insert() expects the 'location' iter to be at the end of the inserted text, rather than the beginning (this is actually the behavior of the default handler). Thus, successive pieces of text were being inserted in the wrong location. The solution, then is to keep `location` where it is (or set it back to the end if it gets invalidated by another insert, such as with adding a checkbox/bullet) and use a separate iter instead.

Fixes #80